taskbar-dnd: Add ghost icon to overflow container
In this cl, we reuse the functionality to create and locate the
ghost icon that is implemented for TaskbarView to the overflow
container. To fulfill that, an abstract class PinnedAppDragHelper
is created to share some logics between the two container views.
Following the composition concept, each of them owns an instance
of the helper to determine the positioning of the ghost icon.

Bug: 447444838
Test: Manual
Flag: com.android.launcher3.enable_taskbar_drag_and_drop
Change-Id: Icc7ee4c3b6e17b9b91d2990a32cc54fbec81ea59