Remove collapse on top resumed from bottom sheet view of widget picker.
The floating sheet view that's desktop specific, closes it. For foldable
on folding the phone, this was collapsing the view but not closing the
activity - so removed that logic as its not necessary there. Also, the
flag for desktop that switches to floating sheet is already released,
so its safe to assume floating sheet is used in desktop.

Bug: 463543893
Test: Cleaned up tests and tried with a foldable
Flag: EXEMPT BUGFIX
Change-Id: I03902fc42a619bf220136f0ccf9e0199332f0126