Fixed bubble bar expanded view overlaps with bubble bar
The overlap was caused by Launcher process sending the bubble bar top
from the bottom of the screen in the middle of the in-app animation,
resulting in an incorrect intermediate translation Y value.

Attempt to send the update after the in-app animation completed led to
the task view in the bubble bar expanded view resizing, which provoked
showing the IME (user closes the IME -> on IME hidden task view is
resized -> IME shown again).

Fixes: 457955659
Test: PersistentBubbleStashControllerTest
Flag: com.android.wm.shell.enable_bubble_bar
Change-Id: I9a2cc6db00f2b8e810c42151dc42ce5fe00a3fc1