Always run the workspace reveal animation.
Before this CL, the Launcher reveal animation only runs as part of
the remote transition that also animates an app going away, or as
part of the recents -> home transitions.

We are introducing the ability for System UI to handle the animation
of an app closing to home in certain scenarios, but we still want
Launcher to run its internal reveal animation. To do so, this CL
makes it so that the reveal animation is run whenever Launcher
was invisible and becomes visible in its NORMAL state.

If Launcher ends up actually handling the whole transition, this
fallback is cancelled and the regular animation runs seamlessly
instead.

Bug: 202516970
Flag: com.android.launcher3.fallback_reveal_animation
Test: manual
Change-Id: I61726545c9eefb61e0037a4b507c04edb62566cd