Stop attempting to go to invalid page on foldable
When trying to go past the last page on foldable, we actually attempt to
go to an invalid page, we should be reaching the else statement. Instead
we reach the else if. On handheld we reach the else statement because
childcount -1 properly accounts for panel count at that point, but in
foldables the logic breaks down because we have one more panel beyond
currentPage

Fix: 466419140
Flag: EXEMPT BUGFIX
Test: verified on device with logs
Change-Id: I9f99952a3a97c1463ba8c6b3c07c57b86bede80b