Fix an issue where taskbar ui state could be out of sync from bubbles
When folding / unfolding bubbles switches between floating & bubble bar
and bubble bar is recreated.

Bubbles could be expanded in bubble bar & taskbarUiState would save
that. Then when bubble UI changes to floating, bubbles may be collapsed.
When you go back to bubble bar UI, taskbarUiState is reused and still
shows that bubble bar is expanded when it isn't.

The fix is to sync it with what BubbleBarView has when taskbarUiStat
is set on the view.

Test: manual - expand bubbles in bubble with camera and another bubble
             - fold the device
             - swipe back to dismiss camera
             - unfold the device
             - tap on bubble bar
             => verify that bubble bar expands
Flag: com.android.launcher3.refactor_taskbar_ui_state
Bug: 463854941
Change-Id: Id4c11cc518e2c01fba42e84410546f04772864ef