Dragging from workspace to taskbar
Has workspace start system drag when taskbar drag and drop is enabled,
so the items can be dragged to taskbar window. The clip data is set to
the dragged item info wrapped in ObjectWrapper, a parcelable IBinder
wrapper around unparcelable item info - within the Launcher process,
binder objects are singletons, so wrapping and unwrapping the
ObjectWrapper within Launcher will return the same backing object.

On taskbar side, adds system drag handler in TaskbarDragController that
starts DragController drag in response to external drag sequence. The
internal drag will start with invisible bitmap (which is not needed,
given the presence of the system drag shadow view). The item info on the
internal drag object is intially set to SystemDragItemInfo, and replaced
with the actual dragged item info in response to drop event.

Bug:468449419
Flag: com.android.launcher3.enable_taskbar_drag_and_drop
Test: Manual

Change-Id: I9711c94570ab7a462a0e97ee3f4b20955ba5e1f1