Schedule swipe to PiP animator txs on RenderThread
SwipePipToHomeAnimator is created and run
by AbsSwipeUpHandler on Launcher's UiThread where
other RecentsView animations also run.

This means SwipePipToHomeAnimator should not independently
create SurfaceControl.Transaction-s to be applied before next vsync.
Instead, use SyncRtSurfaceTransactionApplier to schedule our
transactions to be applied on RenderThread in sync with
all the other draws and buffer updates coming from Launcher.

Since we are animating the surface we should furthermore
pass the leash we are animating into the InteractionJankMonitor
to make sure that jank instrumentation is correct.

Bug: 352743280
Flag: EXEMPT BUGFIX
Test: repro the steps in the BR
Change-Id: If678cd17dfa2dbe5c840485e6bdb43cd4f6c6a4e