Launcher3: Refactor page indicator with arrows and update colors
The page indicator with arrows layout has been refactored to wrap the
arrows and dots in a LinearLayout with a rounded background.
The visibility logic for the arrows has been updated to also control
the visibility of this new container.
Additionally, the page indicator dot color has been changed to
system_accent1_200 for better contrast.

Bug: 456238912
Test: Manual
Flag: EXEMPT BUGFIX
Change-Id: I628f2e605d26c9a7a9b02af95ee071c73c02a4ac