Make folders on taskbar removable with context menu
In desktop devices, there is a case where users can pin a folder
on the taskbar by switching to tablet mode. This wasn't considered
and there wasn't a way to remove that folder in the desktop mode.
This cl extends the usage of the original popup menu for app icons
on the taskbar to apply to folders and add the remove option for
the them.

Bug: 455936052
Test: Manual && TaskbarPopupControllerTest
Flag: EXEMPT BUGFIX
Change-Id: I28a0b354d4613a1c04c35d2a23bee1bb0edfc4b3